+++ packaging/metadata/__init__.py
@@ -1,4 +1,4 @@
-from ._parse import RawMetadata, parse_email, parse_json
+from ._raw import RawMetadata, parse_email, parse_json
 from ._types import DynamicField, Metadata
 
 __all__ = ["DynamicField", "Metadata", "RawMetadata", "parse_email", "parse_json"]
